Refactor page render tests out into separate class
