+++ tests/Cms/Pages/PageCacheTest.php
@@ -2,50 +2,25 @@
 
 namespace Kirby\Cms;
 
-use Kirby\Cache\Value;
-use Kirby\Filesystem\Dir;
 use Kirby\TestCase;
 
 class PageCacheTest extends TestCase
 {
-	public const FIXTURES = __DIR__ . '/fixtures/PageCacheTest';
-	public const TMP      = KIRBY_TMP_DIR . '/Cms.PageCache';
+	public const TMP = KIRBY_TMP_DIR . '/Cms.PageCache';
 
 	public function setUp(): void
 	{
 		$this->app = new App([
 			'roots' => [
-				'index'     => static::TMP,
-				'templates' => static::FIXTURES
+				'index' => static::TMP
 			],
 			'site' => [
 				'children' => [
 					[
-						'slug' => 'default'
+						'slug' => 'one'
 					],
 					[
-						'slug'     => 'expiry',
-						'template' => 'expiry'
-					],
-					[
-						'slug'     => 'disabled',
-						'template' => 'disabled'
-					],
-					[
-						'slug'     => 'dynamic-auth',
-						'template' => 'dynamic'
-					],
-					[
-						'slug'     => 'dynamic-cookie',
-						'template' => 'dynamic'
-					],
-					[
-						'slug'     => 'dynamic-session',
-						'template' => 'dynamic'
-					],
-					[
-						'slug'     => 'dynamic-auth-session',
-						'template' => 'dynamic'
+						'slug' => 'another'
 					]
 				]
 			],
@@ -53,19 +28,6 @@ class PageCacheTest extends TestCase
 				'cache.pages' => true
 			]
 		]);
-
-		Dir::make(static::TMP);
-	}
-
-	public function tearDown(): void
-	{
-		Dir::remove(static::TMP);
-
-		unset(
-			$_COOKIE['foo'],
-			$_COOKIE['kirby_session'],
-			$_SERVER['HTTP_AUTHORIZATION']
-		);
 	}
 
 	public static function requestMethodProvider(): array
@@ -91,7 +53,7 @@ class PageCacheTest extends TestCase
 			]
 		]);
 
-		$this->assertSame($expected, $app->page('default')->isCacheable());
+		$this->assertSame($expected, $app->page('one')->isCacheable());
 	}
 
 	/**
@@ -106,7 +68,7 @@ class PageCacheTest extends TestCase
 			]
 		]);
 
-		$this->assertFalse($app->page('default')->isCacheable());
+		$this->assertFalse($app->page('one')->isCacheable());
 	}
 
 	public function testRequestParams()
@@ -117,7 +79,7 @@ class PageCacheTest extends TestCase
 			]
 		]);
 
-		$this->assertFalse($app->page('default')->isCacheable());
+		$this->assertFalse($app->page('one')->isCacheable());
 	}
 
 	public function testIgnoreId()
@@ -126,14 +88,14 @@ class PageCacheTest extends TestCase
 			'options' => [
 				'cache.pages' => [
 					'ignore' => [
-						'expiry'
+						'another'
 					]
 				]
 			]
 		]);
 
-		$this->assertTrue($app->page('default')->isCacheable());
-		$this->assertFalse($app->page('expiry')->isCacheable());
+		$this->assertTrue($app->page('one')->isCacheable());
+		$this->assertFalse($app->page('another')->isCacheable());
 	}
 
 	public function testIgnoreCallback()
@@ -141,13 +103,13 @@ class PageCacheTest extends TestCase
 		$app = $this->app->clone([
 			'options' => [
 				'cache.pages' => [
-					'ignore' => fn ($page) => $page->id() === 'default'
+					'ignore' => fn ($page) => $page->id() === 'one'
 				]
 			]
 		]);
 
-		$this->assertFalse($app->page('default')->isCacheable());
-		$this->assertTrue($app->page('expiry')->isCacheable());
+		$this->assertFalse($app->page('one')->isCacheable());
+		$this->assertTrue($app->page('another')->isCacheable());
 	}
 
 	public function testDisabledCache()
@@ -159,7 +121,7 @@ class PageCacheTest extends TestCase
 			]
 		]);
 
-		$this->assertFalse($app->page('default')->isCacheable());
+		$this->assertFalse($app->page('one')->isCacheable());
 
 		// deactivate in array
 		$app = $this->app->clone([
@@ -170,158 +132,6 @@ class PageCacheTest extends TestCase
 			]
 		]);
 
-		$this->assertFalse($app->page('default')->isCacheable());
-	}
-
-	public function testRenderCache()
-	{
-		$cache = $this->app->cache('pages');
-		$page  = $this->app->page('default');
-
-		$this->assertNull($cache->retrieve('default.html'));
-
-		$html1 = $page->render();
-		$this->assertStringStartsWith('This is a test:', $html1);
-
-		$value = $cache->retrieve('default.html');
-		$this->assertInstanceOf(Value::class, $value);
-		$this->assertSame($html1, $value->value()['html']);
-		$this->assertNull($value->expires());
-
-		$html2 = $page->render();
-		$this->assertSame($html1, $html2);
-	}
-
-	public function testRenderCacheCustomExpiry()
-	{
-		$cache = $this->app->cache('pages');
-		$page  = $this->app->page('expiry');
-
-		$this->assertNull($cache->retrieve('expiry.html'));
-
-		$time = $page->render();
-
-		$value = $cache->retrieve('expiry.html');
-		$this->assertInstanceOf(Value::class, $value);
-		$this->assertSame($time, $value->value()['html']);
-		$this->assertSame((int)$time, $value->expires());
-	}
-
-	public function testRenderCacheDisabled()
-	{
-		$cache = $this->app->cache('pages');
-		$page  = $this->app->page('disabled');
-
-		$this->assertNull($cache->retrieve('disabled.html'));
-
-		$html1 = $page->render();
-		$this->assertStringStartsWith('This is a test:', $html1);
-
-		$this->assertNull($cache->retrieve('disabled.html'));
-
-		$html2 = $page->render();
-		$this->assertNotSame($html1, $html2);
-	}
-
-	public static function dynamicProvider(): array
-	{
-		return [
-			['dynamic-auth', ['auth']],
-			['dynamic-cookie', ['cookie']],
-			['dynamic-session', ['session']],
-			['dynamic-auth-session', ['auth', 'session']],
-		];
-	}
-
-	/**
-	 * @dataProvider dynamicProvider
-	 */
-	public function testRenderCacheDynamicNonActive(string $slug, array $dynamicElements)
-	{
-		$cache = $this->app->cache('pages');
-		$page  = $this->app->page($slug);
-
-		$this->assertNull($cache->retrieve($slug . '.html'));
-
-		$html1 = $page->render();
-		$this->assertStringStartsWith('This is a test:', $html1);
-
-		$cacheValue = $cache->retrieve($slug . '.html');
-		$this->assertNotNull($cacheValue);
-		$this->assertSame(in_array('auth', $dynamicElements), $cacheValue->value()['usesAuth']);
-		if (in_array('cookie', $dynamicElements)) {
-			$this->assertSame(['foo'], $cacheValue->value()['usesCookies']);
-		} elseif (in_array('session', $dynamicElements)) {
-			$this->assertSame(['kirby_session'], $cacheValue->value()['usesCookies']);
-		} else {
-			$this->assertSame([], $cacheValue->value()['usesCookies']);
-		}
-
-		// reset the Kirby Responder object
-		$this->setUp();
-		$html2 = $page->render();
-		$this->assertSame($html1, $html2);
-	}
-
-	/**
-	 * @dataProvider dynamicProvider
-	 */
-	public function testRenderCacheDynamicActiveOnFirstRender(string $slug, array $dynamicElements)
-	{
-		$_COOKIE['foo'] = $_COOKIE['kirby_session'] = 'bar';
-		$this->app->clone([
-			'server' => [
-				'HTTP_AUTHORIZATION' => 'Bearer brown-bearer'
-			]
-		]);
-
-		$cache = $this->app->cache('pages');
-		$page  = $this->app->page($slug);
-
-		$this->assertNull($cache->retrieve($slug . '.html'));
-
-		$html1 = $page->render();
-		$this->assertStringStartsWith('This is a test:', $html1);
-
-		$cacheValue = $cache->retrieve($slug . '.html');
-		$this->assertNull($cacheValue);
-
-		// reset the Kirby Responder object
-		$this->setUp();
-		$html2 = $page->render();
-		$this->assertNotSame($html1, $html2);
-	}
-
-	/**
-	 * @dataProvider dynamicProvider
-	 */
-	public function testRenderCacheDynamicActiveOnSecondRender(string $slug, array $dynamicElements)
-	{
-		$cache = $this->app->cache('pages');
-		$page  = $this->app->page($slug);
-
-		$this->assertNull($cache->retrieve($slug . '.html'));
-
-		$html1 = $page->render();
-		$this->assertStringStartsWith('This is a test:', $html1);
-
-		$cacheValue = $cache->retrieve($slug . '.html');
-		$this->assertNotNull($cacheValue);
-		$this->assertSame(in_array('auth', $dynamicElements), $cacheValue->value()['usesAuth']);
-		if (in_array('cookie', $dynamicElements)) {
-			$this->assertSame(['foo'], $cacheValue->value()['usesCookies']);
-		} elseif (in_array('session', $dynamicElements)) {
-			$this->assertSame(['kirby_session'], $cacheValue->value()['usesCookies']);
-		} else {
-			$this->assertSame([], $cacheValue->value()['usesCookies']);
-		}
-
-		$_COOKIE['foo'] = $_COOKIE['kirby_session'] = 'bar';
-		$_SERVER['HTTP_AUTHORIZATION'] = 'Bearer brown-bearer';
-
-		// reset the Kirby Responder object
-		$this->setUp();
-		$html2 = $page->render();
-		$this->assertNotSame($html1, $html2);
+		$this->assertFalse($app->page('one')->isCacheable());
 	}
 }
