+++ tests/Cms/Pages/fixtures/PageRenderTest/hook-foo.php
